refactor: fixes and tweaks
