+++ internal/mod.ts
@@ -1,0 +1,9 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+
+/**
+ * Internal utilities for the public API of the Deno Standard Library.
+ *
+ * Note: for internal use only.
+ *
+ * @module
+ */
