+++ internal/testdata/fn.ts
@@ -1,3 +1,4 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { warnDeprecatedApi } from "../warn_deprecated_api.ts";
 
 function fn() {
